live_chess_detection/requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Deep Learning Models
ultralytics>=8.0.0
torchsummary>=1.5.1
timm>=0.9.0
tensorflow>=2.13.0

# Chess Logic
python-chess>=1.9.4
cairosvg>=2.7.1

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Web UI
gradio>=4.0.0
plotly>=5.17.0
dash>=2.14.0

# Experiment Tracking
mlflow>=2.8.0
tensorboard>=2.15.0

# Video Processing
pytube>=15.0.0
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# HTTP and APIs
requests>=2.31.0
aiohttp>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and Logging
tqdm>=4.66.0
colorlog>=6.7.0

# Data Validation
pydantic>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0